+++ crates/vertigo-macro/src/html_parser/node.rs
@@ -162,8 +162,10 @@ fn push_attribute(
 
     let method_str = match name.as_str() {
         "hook_key_down" | "on_blur" | "on_change" | "on_change_file" | "on_click"
-        | "on_dropfile" | "on_input" | "on_key_down" | "on_load" | "on_mouse_down"
-        | "on_mouse_enter" | "on_mouse_leave" | "on_mouse_up" | "on_submit" => &name,
+        | "on_dropfile" | "on_input" | "on_intersect" | "on_key_down" | "on_load"
+        | "on_mouse_down" | "on_mouse_enter" | "on_mouse_leave" | "on_mouse_up" | "on_submit" => {
+            &name
+        }
 
         "form" => "on_submit",
 
@@ -273,6 +275,7 @@ mod tests {
             "on_click",
             "on_change",
             "on_input",
+            "on_intersect",
             "on_blur",
             "on_submit",
             "on_key_down",