+++ crates/vertigo/src/lib.rs
@@ -58,7 +58,7 @@ pub use dom::{
     dom_id::DomId,
     dom_node::DomNode,
     dom_text::DomText,
-    events::{ClickEvent, DropFileEvent, DropFileItem, KeyDownEvent},
+    events::{ClickEvent, DropFileEvent, DropFileItem, IntersectionEvent, KeyDownEvent},
 };
 pub use dom_macro::{AttrGroup, AttrGroupValue, EmbedDom};
 pub use driver_module::{